+++ backend/src/main/java/com/useranalyzer/backend/dto/BarChartData.java
@@ -1,5 +1,6 @@
 package com.useranalyzer.backend.dto;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,6 +11,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class BarChartData {
+    @JsonProperty("xAxis")
     private List<String> xAxis;
     private List<BarSeries> series;