+++ backend/src/main/java/com/useranalyzer/backend/dto/LineChartData.java
@@ -1,5 +1,6 @@
 package com.useranalyzer.backend.dto;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,6 +12,7 @@ import java.util.List;
 @AllArgsConstructor
 public class LineChartData {
     private String name;
+    @JsonProperty("xAxis")
     private List<String> xAxis;
     private List<? extends Number> series;
 }